refactor(crypto): consolidate mint-quote signing in NUT20, mark legacy
cashubtc/nuts#375 defines one mint-quote message for NUT-20 single and NUT-29
batch minting, so the NUT29 module dissolves into NUT20.ts: signMintQuote /
verifyMintQuoteSignature now produce the amended format, and the pre-amendment
concatenation moves to signMintQuoteLegacy / verifyMintQuoteSignatureLegacy.
The transition cleanup is then pure deletion of the Legacy-suffixed pair.
