+++ src/crypto/NUT20.ts
@@ -1,22 +1,66 @@
-import { schnorr } from '@noble/curves/secp256k1.js';
+import { numberToBytesBE } from '@noble/curves/utils.js';
 import { sha256 } from '@noble/hashes/sha2.js';
-import { bytesToHex, hexToBytes } from '@noble/hashes/utils.js';
+import { hexToBytes, utf8ToBytes } from '@noble/hashes/utils.js';
 
+import { Amount } from '../model/Amount';
 import { type SerializedBlindedMessage } from '../model/types';
 
+import { schnorrSignDigest, schnorrVerifyDigest } from './core';
+
+// Domain-separation tag.
+const MINT_QUOTE_SIG_DST = utf8ToBytes('Cashu_MintQuoteSig_v1');
+
+// Canonical minimal BE bytes of a non-negative amount (0 → empty, 1 → 0x01, 256 → 0x0100).
+// Amount.from defensively normalizes a raw JSON number/string (Amount passes through).
+function amountToMinimalBytes(blindedMessage: SerializedBlindedMessage): Uint8Array {
+  const value = Amount.from(blindedMessage.amount).toBigInt();
+  if (value === 0n) return new Uint8Array(0);
+  const hex = value.toString(16);
+  return hexToBytes(hex.length % 2 === 1 ? '0' + hex : hex);
+}
+
 /**
- * Legacy NUT-20 mint-quote signature message: `quote || B_0 || … || B_(n-1)`, hex strings
- * concatenated as UTF-8. cashubtc/nuts#375 replaces this with the domain-separated, length-framed
- * message in `NUT29.ts` for both single (NUT-20) and batch (NUT-29) minting. Kept for mints that
- * predate the amendment — see `wallet/mintCompat.ts`.
+ * Mint-quote signature message per the amended NUT-20 (cashubtc/nuts#375): domain-separated and
+ * length-framed, shared by NUT-20 single and NUT-29 batch minting.
  */
 function constructMessage(quote: string, blindedMessages: SerializedBlindedMessage[]): Uint8Array {
+  // Stream into the digest rather than concat-then-hash: spreading the per-output chunks into
+  // concatBytes(...) would hit V8's argument-count limit on large batches.
+  const transcript = sha256.create();
+  transcript.update(MINT_QUOTE_SIG_DST);
+  const quoteBytes = utf8ToBytes(quote);
+  transcript.update(numberToBytesBE(quoteBytes.length, 4));
+  transcript.update(quoteBytes);
+  for (const blindedMessage of blindedMessages) {
+    const amountBytes = amountToMinimalBytes(blindedMessage);
+    transcript.update(numberToBytesBE(amountBytes.length, 4));
+    transcript.update(amountBytes);
+    const pointBytes = hexToBytes(blindedMessage.B_);
+    transcript.update(numberToBytesBE(pointBytes.length, 4));
+    transcript.update(pointBytes);
+  }
+  return transcript.digest();
+}
+
+/**
+ * Legacy mint-quote signature message: `quote || B_0 || … || B_(n-1)`, hex strings concatenated as
+ * UTF-8. Kept for mints that predate cashubtc/nuts#375 — see `wallet/mintCompat.ts`; delete with
+ * the last `AMENDED_QUOTE_SIG_RELEASES` entry.
+ */
+function constructLegacyMessage(
+  quote: string,
+  blindedMessages: SerializedBlindedMessage[],
+): Uint8Array {
   let message = quote;
   for (const blindedMessage of blindedMessages) {
     message += blindedMessage.B_;
   }
-  const msgbytes = new TextEncoder().encode(message);
-  return sha256(msgbytes);
+  return sha256(utf8ToBytes(message));
+}
+
+// NUT-20 quote pubkeys are compressed 33-byte SEC1 (66 hex chars).
+function isCompressedPubkey(pubkey: string): boolean {
+  return pubkey.length === 66;
 }
 
 export function signMintQuote(
@@ -24,10 +68,7 @@ export function signMintQuote(
   quote: string,
   blindedMessages: SerializedBlindedMessage[],
 ): string {
-  const message = constructMessage(quote, blindedMessages);
-  const privkeyBytes = hexToBytes(privkey);
-  const signature = schnorr.sign(message, privkeyBytes);
-  return bytesToHex(signature);
+  return schnorrSignDigest(constructMessage(quote, blindedMessages), privkey);
 }
 
 export function verifyMintQuoteSignature(
@@ -36,10 +77,24 @@ export function verifyMintQuoteSignature(
   blindedMessages: SerializedBlindedMessage[],
   signature: string,
 ): boolean {
-  const sigbytes = hexToBytes(signature);
-  let pubkeyBytes = hexToBytes(pubkey);
-  if (pubkeyBytes.length !== 33) return false;
-  pubkeyBytes = pubkeyBytes.slice(1);
-  const message = constructMessage(quote, blindedMessages);
-  return schnorr.verify(sigbytes, message, pubkeyBytes);
+  if (!isCompressedPubkey(pubkey)) return false;
+  return schnorrVerifyDigest(signature, constructMessage(quote, blindedMessages), pubkey);
+}
+
+export function signMintQuoteLegacy(
+  privkey: string,
+  quote: string,
+  blindedMessages: SerializedBlindedMessage[],
+): string {
+  return schnorrSignDigest(constructLegacyMessage(quote, blindedMessages), privkey);
+}
+
+export function verifyMintQuoteSignatureLegacy(
+  pubkey: string,
+  quote: string,
+  blindedMessages: SerializedBlindedMessage[],
+  signature: string,
+): boolean {
+  if (!isCompressedPubkey(pubkey)) return false;
+  return schnorrVerifyDigest(signature, constructLegacyMessage(quote, blindedMessages), pubkey);
 }
