+++ test/wallet/wallet-mint.node.test.ts
@@ -18,7 +18,7 @@ import {
 } from '../../src';
 
 import { Bytes, sumProofs } from '../../src/utils';
-import { verifyMintQuoteSignature, verifyBatchMintQuoteSignature } from '../../src/crypto';
+import { verifyMintQuoteSignature, verifyMintQuoteSignatureLegacy } from '../../src/crypto';
 import { AMENDED_QUOTE_SIG_RELEASES } from '../../src/wallet/mintCompat';
 import { hexToBytes } from '@noble/curves/utils.js';
 import { useTestServer, mint, mintUrl, unit, logger, mintInfoResp } from './_setup';
@@ -644,10 +644,10 @@ describe('mint quote signature format gating', () => {
       'signs with the legacy NUT-20 message for %s below %s',
       async (implementation) => {
         const { signature, outputs } = await prepareLocked(`${implementation}/0.0.1`);
-        expect(verifyMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(true);
-        expect(verifyBatchMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(
-          false,
+        expect(verifyMintQuoteSignatureLegacy(pubkey, 'locked-quote', outputs, signature)).toBe(
+          true,
         );
+        expect(verifyMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(false);
       },
     );
 
@@ -655,16 +655,16 @@ describe('mint quote signature format gating', () => {
       'signs with the amended message for %s at %s',
       async (implementation, minVersion) => {
         const { signature, outputs } = await prepareLocked(`${implementation}/${minVersion}`);
-        expect(verifyBatchMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(
-          true,
+        expect(verifyMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(true);
+        expect(verifyMintQuoteSignatureLegacy(pubkey, 'locked-quote', outputs, signature)).toBe(
+          false,
         );
-        expect(verifyMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(false);
       },
     );
 
     test('signs with the amended message for unknown implementations', async () => {
       const { signature, outputs } = await prepareLocked('acme-mint/0.0.1');
-      expect(verifyBatchMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(true);
+      expect(verifyMintQuoteSignature(pubkey, 'locked-quote', outputs, signature)).toBe(true);
     });
   });
 });
